requirements.txt
torch
numpy
scipy
termcolor
tqdm
matplotlib
pyyaml
